Turn unit tests into XCTest which could be run in Xcode or in command line with `swift test`
