+++ tests/LinuxMain.swift
@@ -1,0 +1,1 @@
+fatalError("Run the tests with `swift test --enable-test-discovery`.")
